Create localserver
Local DNS Server: The local DNS server handles requests from the client. It first checks its own RR table for the request. If it’s not found, it asks the authoritative DNS server for the requested hostname/domain. When it gets a response, the local DNS server saves the result/record in its RR table and sends it to the client. Then, it prints out its RR table.